Extract logic for page header
